解决： ./test: line 18: namespace: command not found ./test: line 22: syntax error near unexpected token `(' ./test: line 22: `  typedef decltype(nullptr) nullptr_t;'
原因:https://askubuntu.com/questions/1125126/errors-from-running-code-compiled-by-gcc
cpp is the C pre-processor - it is not a compiler. Its output is a text file containing expanded macros and declarations from any files that your code has #includeed. When you made it executable and tried to run it, your shell made an attempt to interpret this text as shell code - and failed.

To compile your C++ file (and link into an executable) you should use g++ instead:

g++ test.cpp -o test